+++ lib/noted/project/destroy.rb
@@ -9,7 +9,9 @@ class ProjectDestroy < Mutations::Command
   def execute
     p = Project.find(self.project['id'])
 
-    add_error(:current_user, :unauthorized, 'not authorized') unless current_user.can_update?(p)
+    return add_error(:current_user, :unauthorized, 'not authorized') unless current_user.can_update?(p)
+
+    p.updater = current_user
 
     p.destroy
